发光材质

+++ Assets/Scripts/Role/Data/RoleData.cs
@@ -9,7 +9,7 @@ public class RoleData : IcfgId
 {
     public int cfgId { get; set; }
     public string name;
-    public int walkSpeed;
+    public float walkSpeed;
     public int jumpForce;
     public float eatLong;
     public float eatWidth;
@@ -22,6 +22,6 @@ public class RoleData : IcfgId
         this.eatLong = 1;
         this.eatWidth = 1;
         this.interativeRange = 1;
-       
+
     }
 }
